LayoutTests/http/tests/eventloop/perform-microtask-checkpoint-at-end-of-task-expected.txt

A microtask checkpoint should be performed after executing each task.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS logs.join(", ") is "1.task, 2.microtask, 3.microtask, 4.task"
PASS successfullyParsed is true

TEST COMPLETE
